+++ src/luamod.rs
@@ -746,12 +746,10 @@ pub(crate) fn setup() {
 
                 let node = traft::node::global()?;
                 let instance = node.storage.instances.get(&instance_name)?;
-                let cluster_name = node.topology_cache.cluster_name.into();
                 fiber::block_on(rpc::network_call_to_leader(
                     crate::proc_name!(rpc::expel::proc_expel),
                     &rpc::expel::Request {
                         instance_uuid: instance.uuid,
-                        cluster_name,
                         force,
                         timeout,
                     },